fix(bt): Fixed l2cap not reporting stop server completion event(v5.4)

See merge request espressif/esp-idf!38698

+++ components/bt/host/bluedroid/api/include/api/esp_l2cap_bt_api.h
@@ -124,7 +124,7 @@ typedef union {
      */
     struct l2cap_srv_stop_evt_param {
         esp_bt_l2cap_status_t  status;         /*!< status */
-        uint8_t                psm;            /*!< local psm */
+        uint16_t               psm;            /*!< local psm */
     } srv_stop;                                /*!< L2CAP callback param of ESP_BT_L2CAP_SRV_STOP_EVT */
 
     /**
